backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
replicate==0.25.1
werkzeug==3.0.1
gunicorn==21.2.0
